+++ Source/Testably.Abstractions.Testing/FileSystem/FileStreamFactoryMock.cs
@@ -1,6 +1,6 @@
 ﻿using System.IO;
 using Testably.Abstractions.FileSystem;
-using Testably.Abstractions.Testing.Internal;
+using Testably.Abstractions.Testing.Helpers;
 
 namespace Testably.Abstractions.Testing.FileSystem;
