+++ Source/Testably.Abstractions.Testing/FileSystemInitializer/DirectoryCleaner.cs
@@ -2,7 +2,7 @@
 using System.IO;
 using System.Threading;
 using Testably.Abstractions.FileSystem;
-using Testably.Abstractions.Testing.Internal;
+using Testably.Abstractions.Testing.Helpers;
 
 namespace Testably.Abstractions.Testing.FileSystemInitializer;
