+++ Source/Testably.Abstractions.Testing/Helpers/EnumerationOptionsHelper.cs
@@ -2,7 +2,7 @@
 using System.IO;
 using System.IO.Enumeration;
 
-namespace Testably.Abstractions.Testing.Internal;
+namespace Testably.Abstractions.Testing.Helpers;
 
 internal static class EnumerationOptionsHelper
 {
