+++ Source/Testably.Abstractions.Testing/Helpers/FileSystemExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Testably.Abstractions.Testing.Storage;
 
-namespace Testably.Abstractions.Testing.Internal;
+namespace Testably.Abstractions.Testing.Helpers;
 
 internal static class FileSystemExtensions
 {
