+++ Source/Testably.Abstractions.Testing/Polyfills/FileFeatureExtensionMethods.cs
@@ -1,7 +1,7 @@
 ﻿#if !FEATURE_PATH_ADVANCED
 using System.Diagnostics.CodeAnalysis;
 using Testably.Abstractions.FileSystem;
-using Testably.Abstractions.Testing.Internal;
+using Testably.Abstractions.Testing.Helpers;
 
 // ReSharper disable once CheckNamespace
 namespace Testably.Abstractions.Testing;
