+++ Source/Testably.Abstractions.Testing/Storage/InMemoryContainer.cs
@@ -5,7 +5,7 @@ using System.IO;
 using System.Linq;
 using Testably.Abstractions.FileSystem;
 using Testably.Abstractions.Testing.FileSystem;
-using Testably.Abstractions.Testing.Internal;
+using Testably.Abstractions.Testing.Helpers;
 using Testably.Abstractions.TimeSystem;
 using static Testably.Abstractions.Testing.Storage.IStorageContainer;
