+++ Source/Testably.Abstractions.Testing/Storage/InMemoryStorage.cs
@@ -6,7 +6,7 @@ using System.IO;
 using System.Linq;
 using Testably.Abstractions.FileSystem;
 using Testably.Abstractions.Testing.FileSystem;
-using Testably.Abstractions.Testing.Internal;
+using Testably.Abstractions.Testing.Helpers;
 
 namespace Testably.Abstractions.Testing.Storage;
