+++ Source/Testably.Abstractions.Testing/TimeSystem/ThreadMock.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Threading;
-using Testably.Abstractions.Testing.Internal;
+using Testably.Abstractions.Testing.Helpers;
 using Testably.Abstractions.TimeSystem;
 
 namespace Testably.Abstractions.Testing.TimeSystem;
