+++ Tests/Testably.Abstractions.Testing.Tests/Internal/FilePlatformIndependenceExtensionsTests.cs
@@ -1,5 +1,5 @@
 using System.IO;
-using Testably.Abstractions.Testing.Internal;
+using Testably.Abstractions.Testing.Helpers;
 using Testably.Abstractions.Testing.Tests.TestHelpers;
 
 namespace Testably.Abstractions.Testing.Tests.Internal;
