+++ Tests/Testably.Abstractions.Testing.Tests/Internal/PathHelperTests.cs
@@ -1,7 +1,7 @@
 ﻿using Moq;
 using System.IO;
 using Testably.Abstractions.FileSystem;
-using Testably.Abstractions.Testing.Internal;
+using Testably.Abstractions.Testing.Helpers;
 using Testably.Abstractions.Testing.Tests.TestHelpers;
 
 namespace Testably.Abstractions.Testing.Tests.Internal;
