+++ Tests/Testably.Abstractions.Testing.Tests/Storage/InMemoryLocationTests.cs
@@ -1,5 +1,5 @@
 ﻿using System.IO;
-using Testably.Abstractions.Testing.Internal;
+using Testably.Abstractions.Testing.Helpers;
 using Testably.Abstractions.Testing.Storage;
 
 namespace Testably.Abstractions.Testing.Tests.Storage;
