+++ Tests/Testably.Abstractions.Testing.Tests/TestHelpers/LockableContainer.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using System.Security.AccessControl;
 using Testably.Abstractions.FileSystem;
-using Testably.Abstractions.Testing.Internal;
+using Testably.Abstractions.Testing.Helpers;
 using Testably.Abstractions.Testing.Storage;
 using Testably.Abstractions.TimeSystem;
